Fixed the zero-velocity test in the edge/vert collision detection.  Was producing different results with different compilers. Fixed a few warnings.
git-svn-id: https://s.snth.net/svn/neverball/trunk@1347 78b8d119-cf0a-0410-b17c-f493084dd1d7

+++ share/image.c
@@ -114,13 +114,14 @@ static GLuint make_texture(const void *p, int w, int h, int b)
     int k = config_get_d(CONFIG_TEXTURES);
     int W = w;
     int H = h;
-    int max;
+
+    GLint max;
 
     void *q = NULL;
 
     glGetIntegerv(GL_MAX_TEXTURE_SIZE, &max);
 
-    while (w / k > max || h / k > max)
+    while (w / k > (int) max || h / k > (int) max)
         k *= 2;
 
     if (k > 1)
